requirements.txt
allure-behave
allure-python-commons
behave
pylint
selenium
requests
webdriver_manager
pyodbc
pandas